(IMPORTANT) Fix bug limiting randomness in permutation guesses
The default arguments in random_walk were never used, because parser is already giving default values for any argument which is not provided. This means that if the user didn't provide a jump_prob, the default was None, not 0.7. This fix should work functionally but I'm not sure if there's a more beautiful way to do it in Python, so feel free to take a look.